refactor(console): split workspace-detail blocks out of dashboard-overview
console-dashboard-overview.tsx was 1501 lines, 1 over the 1500-line
maintainability guard (test_first_party_code_files_stay_under_line_limit),
reddening CI for every PR against development.

Pure structural move (no JSX/logic/behavior changes): extract the workspace
detail group (TaskDetailsModal, TaskPromptBody, WorkspaceSummary,
WorkspaceRecoveryBlock, OperatorControlsBlock, OperatorControlIcon,
UsageSummaryBlock, ProviderReadinessPreflightBlock, RecoveryCallout,
CoordinationWarningBlock, UsageMetric) into a new sibling
console-dashboard-workspace-detail.tsx, following the existing
console-dashboard-*.tsx pattern. Repoint the one consumer
(console-dashboard.tsx) to import TaskDetailsModal + WorkspaceSummary from
the new file.

overview: 1501 -> 798 lines; new file: 731 lines. Moved and kept component
bodies verified byte-identical against origin/development.

lint, typecheck, build, and the line-limit test all pass.
